feat(my-seat-dang): modify payment url & package name
