+++ src/main/java/com/matdang/seatdang/payment/dto/ReadyRequest.java
@@ -1,4 +1,4 @@
-package com.matdang.seatdang.pay.dto;
+package com.matdang.seatdang.payment.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -10,10 +10,15 @@ import lombok.Data;
 @Builder
 @AllArgsConstructor
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
-public class ApproveRequest {
+public class ReadyRequest {
     private String cid;
-    private String tid;
     private String partnerOrderId;
     private String partnerUserId;
-    private String pgToken;
+    private String itemName;
+    private Integer quantity;
+    private Integer totalAmount;
+    private Integer taxFreeAmount; // 상품 비과세 금액
+    private String approvalUrl;
+    private String cancelUrl;
+    private String failUrl;
 }
