+++ src/main/java/com/matdang/seatdang/payment/dto/ReadyResponse.java
@@ -1,4 +1,4 @@
-package com.matdang.seatdang.pay.dto;
+package com.matdang.seatdang.payment.dto;
 
 import lombok.Data;
