+++ src/main/resources/application-test.yml
@@ -8,3 +8,8 @@ spring:
     username: ${TEST_DB_USERNAME}
     password: ${TEST_DB_PASSWORD}
     driver-class-name: com.mysql.cj.jdbc.Driver
+
+kakao:
+  secret:
+    key: ${KAKAO_PAY_SECRET_KEY}
+cid: ${KAKAO_PAY_CID}
