--- src/main/resources/templates/pay/main.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title> 결제 & 취소</title>
-</head>
-<body>
-<h1> 결제 & 취소</h1>
-<div>
-    <a th:href="@{/readyToKakaoPay/pay/layout}">결제 요청(레이아웃)</a>
-</div>
-<div>
-    <a th:href="@{/readyToKakaoPay/refund/layout}">결제 취소(레이아웃)</a>
-</div>
-<div>
-    <a th:href="@{/readyToKakaoPay/pay/popup}">결제 요청(팝업)</a>
-</div>
-<div>
-    <a th:href="@{/readyToKakaoPay/refund/popup}">결제 취소(팝업)</a>
-</div>
-
-</body>
-</html>
