+++ src/main/resources/templates/payment/main.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title> 결제 & 취소</title>
+</head>
+<body>
+<h1> 결제 & 취소</h1>
+<div>
+    <a th:href="@{/payment/request}">결제 요청(레이아웃)</a>
+</div>
+<div>
+    <a th:href="@{/payment/rufund}">결제 취소(레이아웃)</a>
+</div>
+
+</body>
+</html>
