+++ src/main/resources/templates/payment/ready.html
